+++ .github/workflows/build-and-push.yml
@@ -1,0 +1,87 @@
+name: Build and push to Dockerhub
+
+on:
+  push:
+    branches:
+      - master
+
+jobs:
+  build-mongo:
+    runs-on: ubuntu-20.04
+    timeout-minutes: 30
+    steps:
+      - uses: actions/checkout@v2
+      - name: Set up QEMU
+        uses: docker/setup-qemu-action@v1
+      - name: Set up Docker Buildx
+        id: buildx
+        uses: docker/setup-buildx-action@v1
+      - name: Login to Docker Hub
+        uses: docker/login-action@v1
+        with:
+          username: ${{ secrets.DOCKERHUB_USERNAME }}
+          password: ${{ secrets.DOCKERHUB_TOKEN }}
+      - name: Create mongo 4 container image
+        run: docker buildx build -f dbs/mongo/4.0/Dockerfile -t metabase/qa-databases:mongo-sample-4 --push --compress --no-cache --platform linux/arm64/v8,linux/amd64 .
+      - name: Create mongo 5 container image
+        run: docker buildx build -f dbs/mongo/5.0/Dockerfile -t metabase/qa-databases:mongo-sample-5 --push --compress --no-cache --platform linux/arm64/v8,linux/amd64 .
+    
+  build-mssql:
+    runs-on: ubuntu-20.04
+    timeout-minutes: 30
+    steps:
+      - uses: actions/checkout@v2
+      - name: Set up QEMU
+        uses: docker/setup-qemu-action@v1
+      - name: Set up Docker Buildx
+        id: buildx
+        uses: docker/setup-buildx-action@v1
+      - name: Login to Docker Hub
+        uses: docker/login-action@v1
+        with:
+          username: ${{ secrets.DOCKERHUB_USERNAME }}
+          password: ${{ secrets.DOCKERHUB_TOKEN }}
+      # - name: Create mssql 2017 image
+      #   run: docker buildx build -f dbs/mssql/2017/Dockerfile -t metabase/qa-databases:mssql-sample-2017 --push --compress --no-cache .
+      - name: Create mssql 2019 image
+        run: docker buildx build -f dbs/mssql/2019/Dockerfile -t metabase/qa-databases:mssql-sample-2019 --push --compress --no-cache .
+
+  build-mysql:
+    runs-on: ubuntu-20.04
+    timeout-minutes: 30
+    steps:
+      - uses: actions/checkout@v2
+      - name: Set up QEMU
+        uses: docker/setup-qemu-action@v1
+      - name: Set up Docker Buildx
+        id: buildx
+        uses: docker/setup-buildx-action@v1
+      - name: Login to Docker Hub
+        uses: docker/login-action@v1
+        with:
+          username: ${{ secrets.DOCKERHUB_USERNAME }}
+          password: ${{ secrets.DOCKERHUB_TOKEN }}
+      - name: Create mysql 5.7 image
+        run: docker buildx build -f dbs/mysql/5.7/Dockerfile -t metabase/qa-databases:mysql-sample-5.7 --push --compress --no-cache .
+      - name: Create mysql 8 image
+        run: docker buildx build -f dbs/mysql/8/Dockerfile -t metabase/qa-databases:mysql-sample-8 --push --compress --no-cache .
+
+  build-postgres:
+    runs-on: ubuntu-20.04
+    timeout-minutes: 30
+    steps:
+      - uses: actions/checkout@v2
+      - name: Set up QEMU
+        uses: docker/setup-qemu-action@v1
+      - name: Set up Docker Buildx
+        id: buildx
+        uses: docker/setup-buildx-action@v1
+      - name: Login to Docker Hub
+        uses: docker/login-action@v1
+        with:
+          username: ${{ secrets.DOCKERHUB_USERNAME }}
+          password: ${{ secrets.DOCKERHUB_TOKEN }}
+      - name: Create postgres 9 image
+        run: docker buildx build -f dbs/postgres/9/Dockerfile -t metabase/qa-databases:postgres-sample-9 --push --compress --no-cache --platform linux/arm64/v8,linux/amd64 .
+      - name: Create postgres 14 image
+        run: docker buildx build -f dbs/postgres/14/Dockerfile -t metabase/qa-databases:postgres-sample-14 --push --compress --no-cache --platform linux/arm64/v8,linux/amd64 .
