+++ dbs/mongo/5.0/Dockerfile
@@ -4,8 +4,7 @@ FROM mongo:5.0
 ENV MONGO_INITDB_ROOT_USERNAME=metabase MONGO_INITDB_ROOT_PASSWORD=metasample123 MONGO_INITDB_DATABASE=sample
 
 # We copy in our initialization data
-COPY data/*.tsv /docker-entrypoint-initdb.d/
-COPY data/load.sh /docker-entrypoint-initdb.d/
+COPY dbs/mongo/data/*.tsv dbs/mongo/data/load.sh /docker-entrypoint-initdb.d/
 
 # Copy mongo server and client certificates. The client certificates are stored
 # in the image so that they can be pulled from them for local testing.
