+++ dbs/mssql/2019/Dockerfile
@@ -2,7 +2,7 @@ FROM mcr.microsoft.com/mssql/server:2019-latest
 
 ENV ACCEPT_EULA=Y SA_PASSWORD=Metasample123 MSSQL_PID=Standard
 
-COPY sample_data.sql.gz /home/
+COPY dbs/mssql/data/sample_data.sql.gz /home/
 USER root
 RUN gzip -d /home/sample_data.sql.gz
 USER mssql
